Make javascript testable
The goal here is to have the tests be able to test the javascript code
which is used in the page.

In order to do that we need to extract the javascript out to a separate
file so that the the QUnit tests can test it.

This approach follows the advice in the qunit docs:
https://qunitjs.com/intro/#make-things-testable

...but modified to allow running via the command line (and therefore on
Travis) by following this stackoverflow post:
https://stackoverflow.com/a/51861149