Remove trailing and leading newlines from wordlist
For long word lists it's useful to be able to break the list up into
sub-sections with line breaks.

Previously if there was a line break before or after a word, this would
be rendered in the bingo card.

This change preserves internal line breaks, but removes leading and
trailing breaks.

Note that this handles multiple line breaks in a row.